pyproject.toml
[project]
name = "git-summarizer"
version = "0.1.0"
description = "AI-powered Git progress reports and commit message generator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "InquirerPy>=0.3.4",
    "httpx>=0.27.0",
]

[project.scripts]
git-summarizer = "git_summarizer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/git_summarizer"]

[tool.hatch.build.targets.sdist]
include = ["src/git_summarizer"]
